back-end/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-cors-headers==3.7.0
django-sslserver==0.22
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
idna==2.10
in-place==0.5.0
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
unicodecsv==0.14.1
urllib3==1.26.3
